body_behavior: Implement all head modes
